RAG Assessment/requirements.txt
fastapi[all]>=0.109.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-text-splitters>=1.0.0
openai>=1.10.0
faiss-cpu>=1.8.0
azure-search-documents>=11.4.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pypdf2>=3.0.1
pdfplumber>=0.10.3
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
redis>=5.0.1
pytest>=7.4.4
httpx>=0.26.0
